fix: rename the network module to remote
