+++ app/src/main/java/com/movie_app_task/feature/movie_list/data/repository/remote/MovieRemoteDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package com.movie_app_task.feature.movie_list.data.repository.remote
 
 import com.movie_app_task.common.Constants
-import com.movie_app_task.common.domain.remote.INetworkProvider
+import com.movie_app_task.common.domain.repository.remote.INetworkProvider
 import com.movie_app_task.feature.movie_list.data.models.dto.MoviesResponseDto
 import com.movie_app_task.feature.movie_list.domain.repository.remote.MovieRemoteDataSource
